requirements.txt
numpy
pandas
python-dotenv
nltk
scikit-learn
seaborn
matplotlib
#-e .